move files to diff dir
