+++ packages/react-core/src/hooks/external-store/useTrackedDictionaryResolver.ts
@@ -1,28 +1,27 @@
-import { getDictionaryListenerKey, getI18nConfig } from 'gt-i18n/internal';
-import { useShouldTranslate } from '../../hooks/utils';
-import type {
-  DictionaryLookup,
-  DictionaryObjectSnapshot,
-  StoreListener,
-} from '../storeTypes';
-import { useDictionariesSnapshot, useI18nStore } from '../useI18nStore';
+import { useCallback, useEffect, useRef } from 'react';
 import {
-  type RefObject,
-  useCallback,
-  useEffect,
-  useRef,
-  useSyncExternalStore,
-} from 'react';
+  useDictionariesSnapshot,
+  useI18nStore,
+} from '../../i18n-store/useI18nStore';
+import type {
+  DictionaryEntrySnapshot,
+  DictionaryLookup,
+  StoreListener,
+} from '../../i18n-store/storeTypes';
+import { getDictionaryListenerKey, getI18nConfig } from 'gt-i18n/internal';
+import { useSyncExternalStore } from 'react';
+import type { RefObject } from 'react';
+import { useShouldTranslate } from '../utils';
 
-export type TrackedDictionaryObjResolver = (
+export type TrackedDictionaryEntryResolver = (
   lookup: DictionaryLookup
-) => DictionaryObjectSnapshot;
+) => DictionaryEntrySnapshot;
 
-export type OnMissingDictionaryObj = (lookup: DictionaryLookup) => void;
+export type OnMissingDictionaryEntry = (lookup: DictionaryLookup) => void;
 
-export function useTrackedDictionaryObjResolver(
-  onMissingDictionaryObj: OnMissingDictionaryObj = () => {}
-): TrackedDictionaryObjResolver {
+export function useTrackedDictionaryResolver(
+  onMissingDictionaryEntry: OnMissingDictionaryEntry = () => {}
+): TrackedDictionaryEntryResolver {
   const dictionariesSnapshot = useDictionariesSnapshot();
   const i18nStore = useI18nStore();
   const devHotReloadEnabled = getI18nConfig().isDevHotReloadEnabled();
@@ -44,7 +43,7 @@ export function useTrackedDictionaryObjResolver(
       return;
     }
     pendingLookups.forEach((lookup) => {
-      i18nStore.translateDictionaryObject(lookup);
+      i18nStore.translateDictionaryEntry(lookup);
     });
   }, [i18nStore, pendingLookups, shouldTranslate, devHotReloadEnabled]);
 
@@ -58,25 +57,25 @@ export function useTrackedDictionaryObjResolver(
       }
 
       // Resolve the dictionary entry from the store
-      const dictionaryObject = i18nStore.getDictionaryObjectSnapshot(
+      const dictionaryEntry = i18nStore.getDictionaryEntrySnapshot(
         lookup,
         dictionariesSnapshot
       );
 
       // Hot reload
-      if (dictionaryObject == null && devHotReloadEnabled) {
+      if (dictionaryEntry == null && devHotReloadEnabled) {
         pendingLookups.set(lookupKey, lookup);
-        onMissingDictionaryObj(lookup);
+        onMissingDictionaryEntry(lookup);
       }
 
-      return dictionaryObject;
+      return dictionaryEntry;
     },
     [
       i18nStore,
       dictionariesSnapshot,
       devHotReloadEnabled,
-      pendingLookups,
-      onMissingDictionaryObj,
+      pendingLookups, // TODO: maybe should use a wrapper or smth so we can cache cb
+      onMissingDictionaryEntry,
     ]
   );
 }
@@ -87,7 +86,7 @@ function useSubscribeToLookups(trackedKeysRef: RefObject<Set<string> | null>) {
   const i18nStore = useI18nStore();
   const subscribe = useCallback(
     (listener: StoreListener) => {
-      return i18nStore.subscribeToDictionaryObjectEvents((lookup) => {
+      return i18nStore.subscribeToDictionaryEntryEvents((lookup) => {
         const key = getDictionaryListenerKey(lookup);
         if (!trackedKeysRef.current!.has(key)) return;
         versionRef.current++;
